requirements.txt
streamlit>=1.22.0
pandasql  # this is for examples/ only
st-gsheets-connection